requirements.txt
aniso8601==9.0.1
asyncpg==0.25.0
click==8.0.4
Flask==2.0.3
Flask-RESTful==0.3.9
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
